converted most of the "float" to "double", "Float" to "Double", and "clsMutableFloat" to "clsMutableDouble"

+++ DecisionUnitInterface/src/decisionunit/itf/sensors/clsEnergyConsumption.java
@@ -11,7 +11,7 @@ package decisionunit.itf.sensors;
  */
 public class clsEnergyConsumption extends clsSensorHomeostasis {
 
-	public float mrEnergy;
+	public double mrEnergy;
 	
 	@Override
 	public String logXML() {
